requirements.txt
pandas>=1.3.1
lasio>=0.29
streamlit
altair
Image
StringIO
matplotlib.pyplot